Refactor main layout and isolate components for more granular control over mounting and unmounting
